backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
google-generativeai==0.7.2
httpx==0.27.0
python-multipart==0.0.9
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
cyvcf2==0.30.28
aiofiles==23.2.1
tenacity==8.3.0
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0
